#92 Rename view filenames to uppercase
